Add cast to size result.

+++ core/mem/mem.c
@@ -81,7 +81,7 @@ void core_free(void *buf)
 size_t core_size(void *buf)
 {
     SEGMENT sizeSegs = *((SEGMENT *)(buf - sizeof(SEGMENT)));
-    return (sizeSegs * SEGMENT_SIZE) - sizeof(SEGMENT);
+    return ((size_t)sizeSegs * SEGMENT_SIZE) - sizeof(SEGMENT);
 }
 
 size_t core_avail(MEM_TYPE memtype)
